- Changed subroutes' signatures to struct. - Moved subroutes structs to the ListRoute scene.
