+++ Tripper/Scenes/List Route/ListRouteModels.swift
@@ -12,7 +12,31 @@
 
 import UIKit
 
+protocol Subroute {
+    var title: String { get }
+    var timeInSeconds: Int { get }
+}
+
 enum ListRoute {
+    
+    struct Staying: Subroute {
+        var title: String
+        var timeInSeconds: Int
+        var description: String
+    }
+    
+    struct InRoad: Subroute {
+        var title: String
+        var timeInSeconds: Int
+        var metres: Int
+        
+        init(title: String = "In Road", seconds: Int, metres: Int) {
+            self.title = title
+            timeInSeconds = seconds
+            self.metres = metres
+        }
+    }
+    
     // MARK: Use Cases
     
     enum FetchData {
