--- Tripper/Scenes/List Route/Subroutes/Subroute.swift
@@ -1,14 +1,0 @@
-//
-//  Subroute.swift
-//  Tripper
-//
-//  Created by Denis Cherniy on 05.02.2020.
-//  Copyright © 2020 Denis Cherniy. All rights reserved.
-//
-
-import Foundation
-
-protocol Subroute {
-    var title: String { get }
-    var timeInSeconds: Int { get }
-}
